+++ source/language_service/src/test_utils.rs
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+pub(super) mod openqasm;
+
 use std::sync::Arc;
 
 use crate::compilation::{Compilation, CompilationKind};
@@ -396,55 +398,4 @@ fn target_offsets_to_spans(target_offsets: &[u32]) -> Vec<Span> {
         });
     }
     spans
-}
-
-pub(crate) mod openqasm {
-    use super::get_sources_and_markers;
-    use crate::Compilation;
-    use qsc::{
-        PackageType,
-        line_column::{Position, Range},
-        location::Location,
-    };
-    use std::sync::Arc;
-
-    fn compile_project_with_markers_cursor_optional(
-        sources_with_markers: &[(&str, &str)],
-    ) -> (Compilation, Option<(String, Position)>, Vec<Location>) {
-        let (sources, cursor_location, target_spans) =
-            get_sources_and_markers(sources_with_markers);
-
-        (
-            Compilation::new_qasm(
-                PackageType::Lib,
-                sources,
-                vec![],
-                &Arc::from("test project"),
-            ),
-            cursor_location,
-            target_spans,
-        )
-    }
-
-    pub(crate) fn compile_project_with_markers(
-        sources_with_markers: &[(&str, &str)],
-    ) -> (Compilation, String, Position, Vec<Location>) {
-        let (compilation, cursor_location, target_spans) =
-            compile_project_with_markers_cursor_optional(sources_with_markers);
-
-        let (cursor_uri, cursor_offset) =
-            cursor_location.expect("input string should have a cursor marker");
-
-        (compilation, cursor_uri, cursor_offset, target_spans)
-    }
-
-    pub fn compile_with_markers(source_with_markers: &str) -> (Compilation, Position, Vec<Range>) {
-        let (compilation, _, cursor_offset, target_spans) =
-            compile_project_with_markers(&[("<source>", source_with_markers)]);
-        (
-            compilation,
-            cursor_offset,
-            target_spans.iter().map(|l| l.range).collect(),
-        )
-    }
 }
