+++ source/language_service/src/test_utils/openqasm.rs
@@ -1,0 +1,50 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+use super::get_sources_and_markers;
+use crate::Compilation;
+use qsc::{
+    PackageType,
+    line_column::{Position, Range},
+    location::Location,
+};
+use std::sync::Arc;
+
+fn compile_project_with_markers_cursor_optional(
+    sources_with_markers: &[(&str, &str)],
+) -> (Compilation, Option<(String, Position)>, Vec<Location>) {
+    let (sources, cursor_location, target_spans) = get_sources_and_markers(sources_with_markers);
+
+    (
+        Compilation::new_qasm(
+            PackageType::Lib,
+            sources,
+            vec![],
+            &Arc::from("test project"),
+        ),
+        cursor_location,
+        target_spans,
+    )
+}
+
+pub(crate) fn compile_project_with_markers(
+    sources_with_markers: &[(&str, &str)],
+) -> (Compilation, String, Position, Vec<Location>) {
+    let (compilation, cursor_location, target_spans) =
+        compile_project_with_markers_cursor_optional(sources_with_markers);
+
+    let (cursor_uri, cursor_offset) =
+        cursor_location.expect("input string should have a cursor marker");
+
+    (compilation, cursor_uri, cursor_offset, target_spans)
+}
+
+pub fn compile_with_markers(source_with_markers: &str) -> (Compilation, Position, Vec<Range>) {
+    let (compilation, _, cursor_offset, target_spans) =
+        compile_project_with_markers(&[("<source>", source_with_markers)]);
+    (
+        compilation,
+        cursor_offset,
+        target_spans.iter().map(|l| l.range).collect(),
+    )
+}
